Design a stack that returns a minimum element without using an auxiliary stack in Python
Suppose we will push a number value into a stack with a minimum number, min. If the value is greater than or equal to the min, it is pushed directly into the stack. If it is less than min, push 2×value-min, and update min as a value since a new minimum number is pushed. How about to pop? We pop it directly if the top of the stack (it is denoted as top) is greater than or equal to min. Otherwise, the number top is not the real pushed number. The real pushed number is stored as min. After the current minimum number is popped, we need to restore the previous minimum number, 2×min-top.

Now let’s demonstrate its correctness of this solution. Since the value is greater than or equal to min, it is pushed into stack direct without updating min. Therefore, when we find that the top of the stack is greater than or equal to min, we can pop directly without updating min. However, if we find the value is less than min, push 2×value-min. We should notice that 2×value-min should be less than the value. Then we update the current min as value. Therefore, the new top of the stack is less than the current min. Therefore, when we find that the top of the stack is less than min, the real top (real pushed number value) is stored in min. After we pop the top of the stack, we have to restore the previous minimum number. Since top=2×value-previous-min and value is current min, previous min is 2×current-min-top.