Create Find duplicate rows in a binary matrix.py
(Using Trie)
The idea is to insert each row of the given binary matrix into a binary Trie. The alphabet size of a binary trie is only limited to Boolean numbers (0 and 1). If a row is seen before (i.e., it is already present in the Trie), report it as duplicate.